Making sure the team mock works correctly
