+++ Frontend/src/shared/services/mock/Job.service.js
@@ -1,16 +1,24 @@
+const mockJobs = [
+  {
+    id: 'f7f7',
+    forPart: 'geneMapper',
+    status: 'DONE',
+  },
+  {
+    id: 'f3f6',
+    forPart: 'geneCruncher',
+    status: 'DONE',
+  },
+];
+
 const JobService = {
   async getJobs() {
-    // TODO: add mock data
-    return [];
+    return mockJobs;
   },
 
   async getTeamJobs(teamId, forPart) {
     const allJobs = await JobService.getJobs();
-    // pretty random, this is just mocking
-    if (forPart === 'geneMapper') {
-      return allJobs.slice(teamId, allJobs.length / 2);
-    }
-    return allJobs.slice(allJobs.length / 2, allJobs.length - teamId);
+    return allJobs.filter((elem) => elem.forPart === forPart);
   },
 };
